feat: move to eslint flat config
