--- .eslintignore
@@ -1,19 +1,0 @@
-#
-# Directories
-#
-.vscode/
-.history/
-dist/
-node_modules/
-src/test/__fixtures/
-
-#
-# Files
-#
-CHANGELOG.md
-package*.json
-
-# Enable config file linting except for copyright file
-!.*.js
-!.*.cjs
-.copyright.js
