+++ package.json
@@ -48,6 +48,7 @@
   "devDependencies": {
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
+    "@eslint-community/eslint-plugin-eslint-comments": "^4.3.0",
     "@ibm/telemetry-attributes-js": "^3.2.0",
     "@ibm/telemetry-config-schema": "^1.3.0",
     "@opentelemetry/api": "^1.8.0",
@@ -60,15 +61,12 @@
     "@types/lodash": "^4.17.1",
     "@types/node": "^20.12.11",
     "@types/semver": "^7.5.8",
-    "@typescript-eslint/eslint-plugin": "^7.8.0",
-    "@typescript-eslint/parser": "^7.8.0",
     "@vitest/coverage-v8": "^1.6.0",
     "ajv": "^8.13.0",
     "ci-info": "^4.0.0",
     "commander": "^12.0.0",
     "esbuild": "^0.21.1",
     "eslint": "^8.57.0",
-    "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsdoc": "^48.2.4",
     "eslint-plugin-n": "^16.6.2",
@@ -76,6 +74,7 @@
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-simple-import-sort": "^12.1.0",
     "eslint-plugin-vitest": "^0.4.1",
+    "globals": "^15.3.0",
     "husky": "^9.0.11",
     "js-yaml": "^4.1.0",
     "lint-staged": "^15.2.2",
@@ -87,6 +86,7 @@
     "semver": "^7.6.2",
     "ts-json-schema-generator": "^1.5.1",
     "typescript": "^5.4.5",
+    "typescript-eslint": "^7.10.0",
     "vite": "^5.2.11",
     "vitest": "^1.6.0"
   }
