+++ scripts/lint
@@ -6,7 +6,7 @@ echo "Running prettier"
 npx prettier --check .
 
 echo "Running eslint"
-npx eslint --max-warnings=0 --ext .js,.cjs,.mjs,.jsx,.ts,.tsx .
+npx eslint --max-warnings=0 .
 
 echo "Running package.json formatter"
 node scripts/format-package-json.js
